2024-qunicorn-paper/grover-3-sat/requirements.txt
amazon-braket-sdk
numpy
requests